Move event data into sumodules
No logical changes, just a move.
