+++ src/client_event.rs
@@ -1,17 +1,10 @@
-use std::{
-    any::{self, TypeId},
-    io::Cursor,
-    mem,
-};
+mod client_event_data;
+
+use std::io::Cursor;
 
 use bevy::{
-    ecs::{
-        component::{ComponentId, Components},
-        entity::MapEntities,
-        event::ManualEventReader,
-    },
+    ecs::{entity::MapEntities, event::ManualEventReader},
     prelude::*,
-    ptr::{Ptr, PtrMut},
 };
 use bincode::{DefaultOptions, Options};
 use serde::{de::DeserializeOwned, Serialize};
@@ -26,6 +19,7 @@ use crate::{
     },
     server::{replicon_server::RepliconServer, ServerSet},
 };
+use client_event_data::ClientEventData;
 
 /// An extension trait for [`App`] for creating client events.
 pub trait ClientEventAppExt {
@@ -196,10 +190,10 @@ impl ClientEventPlugin {
                             // SAFETY: both resources mutably borrowed uniquely.
                             let (events, reader) = unsafe {
                                 let events = world_cell
-                                    .get_resource_by_id(event_data.events_id)
+                                    .get_resource_by_id(event_data.events_id())
                                     .expect("events shouldn't be removed");
                                 let reader = world_cell
-                                    .get_resource_mut_by_id(event_data.reader_id)
+                                    .get_resource_mut_by_id(event_data.reader_id())
                                     .expect("event reader shouldn't be removed");
                                 (events, reader)
                             };
@@ -230,7 +224,7 @@ impl ClientEventPlugin {
 
                     for event_data in &event_registry.0 {
                         let client_events = world
-                            .get_resource_mut_by_id(event_data.client_events_id)
+                            .get_resource_mut_by_id(event_data.client_events_id())
                             .expect("client events shouldn't be removed");
 
                         // SAFETY: passed pointer was obtained using this event data.
@@ -250,10 +244,10 @@ impl ClientEventPlugin {
                 // SAFETY: both resources mutably borrowed uniquely.
                 let (client_events, events) = unsafe {
                     let client_events = world_cell
-                        .get_resource_mut_by_id(event_data.client_events_id)
+                        .get_resource_mut_by_id(event_data.client_events_id())
                         .expect("client events shouldn't be removed");
                     let events = world_cell
-                        .get_resource_mut_by_id(event_data.events_id)
+                        .get_resource_mut_by_id(event_data.events_id())
                         .expect("events shouldn't be removed");
                     (client_events, events)
                 };
@@ -270,7 +264,7 @@ impl ClientEventPlugin {
         world.resource_scope(|world, event_registry: Mut<ClientEventRegistry>| {
             for event_data in &event_registry.0 {
                 let events = world
-                    .get_resource_mut_by_id(event_data.events_id)
+                    .get_resource_mut_by_id(event_data.events_id())
                     .expect("events shouldn't be removed");
 
                 // SAFETY: passed pointer was obtained using this event data.
@@ -283,173 +277,6 @@ impl ClientEventPlugin {
 /// Registered client events.
 #[derive(Resource, Default)]
 struct ClientEventRegistry(Vec<ClientEventData>);
-
-/// Type-erased functions and metadata for a registered client event.
-///
-/// Needed so events of different types can be processed together.
-struct ClientEventData {
-    type_id: TypeId,
-    type_name: &'static str,
-
-    /// ID of [`Events<E>`] resource.
-    events_id: ComponentId,
-
-    /// ID of [`ClientEventReader<E>`] resource.
-    reader_id: ComponentId,
-
-    /// ID of [`Events<ToClients<E>>`] resource.
-    client_events_id: ComponentId,
-
-    /// Used channel.
-    channel_id: u8,
-
-    send: SendFn,
-    receive: ReceiveFn,
-    resend_locally: ResendLocallyFn,
-    reset: ResetFn,
-    serialize: unsafe fn(),
-    deserialize: unsafe fn(),
-}
-
-impl ClientEventData {
-    fn new<E: Event>(
-        components: &Components,
-        channel_id: u8,
-        serialize: SerializeFn<E>,
-        deserialize: DeserializeFn<E>,
-    ) -> Self {
-        let events_id = components.resource_id::<Events<E>>().unwrap_or_else(|| {
-            panic!(
-                "event `{}` should be previously registered",
-                any::type_name::<E>()
-            )
-        });
-        let client_events_id = components
-            .resource_id::<Events<FromClient<E>>>()
-            .unwrap_or_else(|| {
-                panic!(
-                    "event `{}` should be previously registered",
-                    any::type_name::<FromClient<E>>()
-                )
-            });
-        let reader_id = components
-            .resource_id::<ClientEventReader<E>>()
-            .unwrap_or_else(|| {
-                panic!(
-                    "resource `{}` should be previously inserted",
-                    any::type_name::<ClientEventReader<E>>()
-                )
-            });
-
-        // SAFETY: these functions won't be called until the type is restored.
-        Self {
-            type_id: TypeId::of::<E>(),
-            type_name: any::type_name::<E>(),
-            events_id,
-            reader_id,
-            client_events_id,
-            channel_id,
-            send: send::<E>,
-            receive: receive::<E>,
-            resend_locally: resend_locally::<E>,
-            reset: reset::<E>,
-            serialize: unsafe { mem::transmute(serialize) },
-            deserialize: unsafe { mem::transmute(deserialize) },
-        }
-    }
-
-    /// Sends an event to the server.
-    ///
-    /// # Safety
-    ///
-    /// The caller must ensure that `events` is [`Events<E>`], `reader` is [`ClientEventReader<E>`]
-    /// and this instance was created for `E`.
-    unsafe fn send(
-        &self,
-        ctx: &mut ClientSendCtx,
-        events: &Ptr,
-        reader: PtrMut,
-        client: &mut RepliconClient,
-    ) {
-        (self.send)(self, ctx, events, reader, client);
-    }
-
-    /// Receives an event from a client.
-    ///
-    /// # Safety
-    ///
-    /// The caller must ensure that `events` is [`Events<FromClient<E>>`]
-    /// and this instance was created for `E`.
-    unsafe fn receive(
-        &self,
-        ctx: &mut ServerReceiveCtx,
-        client_events: PtrMut,
-        server: &mut RepliconServer,
-    ) {
-        (self.receive)(self, ctx, client_events, server);
-    }
-
-    /// Drains events `E` and re-emits them as [`FromClient<E>`].
-    ///
-    /// # Safety
-    ///
-    /// The caller must ensure that `events` is [`Events<E>`], `client_events` is [`Events<FromClient<E>>`]
-    /// and this instance was created for `E`.
-    unsafe fn resend_locally(&self, client_events: PtrMut, events: PtrMut) {
-        (self.resend_locally)(client_events, events);
-    }
-
-    /// Drains all events.
-    ///
-    /// # Safety
-    ///
-    /// The caller must ensure that `events` is [`Events<E>`]
-    /// and this instance was created for `E`.
-    unsafe fn reset(&self, events: PtrMut) {
-        (self.reset)(events);
-    }
-
-    /// Serializes an event into a cursor.
-    ///
-    /// # Safety
-    ///
-    /// The caller must ensure that this instance was created for `E`.
-    unsafe fn serialize<E: Event>(
-        &self,
-        ctx: &mut ClientSendCtx,
-        event: &E,
-        cursor: &mut Cursor<Vec<u8>>,
-    ) -> bincode::Result<()> {
-        self.check_type::<E>();
-        let serialize: SerializeFn<E> = std::mem::transmute(self.serialize);
-        (serialize)(ctx, event, cursor)
-    }
-
-    /// Deserializes an event into a cursor.
-    ///
-    /// # Safety
-    ///
-    /// The caller must ensure that this instance was created for `E`.
-    unsafe fn deserialize<E: Event>(
-        &self,
-        ctx: &mut ServerReceiveCtx,
-        cursor: &mut Cursor<&[u8]>,
-    ) -> bincode::Result<E> {
-        self.check_type::<E>();
-        let deserialize: DeserializeFn<E> = std::mem::transmute(self.deserialize);
-        (deserialize)(ctx, cursor)
-    }
-
-    fn check_type<C: Event>(&self) {
-        debug_assert_eq!(
-            self.type_id,
-            TypeId::of::<C>(),
-            "trying to call event functions with {}, but they were created with {}",
-            any::type_name::<C>(),
-            self.type_name,
-        );
-    }
-}
 
 /// Tracks read events for [`ClientEventPlugin::send`].
 ///
@@ -464,103 +291,11 @@ impl<E: Event> FromWorld for ClientEventReader<E> {
     }
 }
 
-/// Signature of client event sending functions.
-type SendFn = unsafe fn(&ClientEventData, &mut ClientSendCtx, &Ptr, PtrMut, &mut RepliconClient);
-
-/// Signature of client event receiving functions.
-type ReceiveFn = unsafe fn(&ClientEventData, &mut ServerReceiveCtx, PtrMut, &mut RepliconServer);
-
-/// Signature of client event resending functions.
-type ResendLocallyFn = unsafe fn(PtrMut, PtrMut);
-
-/// Signature of client event reset functions.
-type ResetFn = unsafe fn(PtrMut);
-
 /// Signature of client event serialization functions.
 pub type SerializeFn<E> = fn(&mut ClientSendCtx, &E, &mut Cursor<Vec<u8>>) -> bincode::Result<()>;
 
 /// Signature of client event deserialization functions.
 pub type DeserializeFn<E> = fn(&mut ServerReceiveCtx, &mut Cursor<&[u8]>) -> bincode::Result<E>;
-
-/// Typed version of [`ClientEvent::send`].
-///
-/// # Safety
-///
-/// The caller must ensure that `events` is [`Events<FromClient<E>>`], `reader` is [`ClientEventReader<E>`],
-/// and `event_data` was created for `E`.
-unsafe fn send<E: Event>(
-    event_data: &ClientEventData,
-    ctx: &mut ClientSendCtx,
-    events: &Ptr,
-    reader: PtrMut,
-    client: &mut RepliconClient,
-) {
-    let reader: &mut ClientEventReader<E> = reader.deref_mut();
-    for event in reader.read(events.deref()) {
-        let mut cursor = Default::default();
-        event_data
-            .serialize::<E>(ctx, event, &mut cursor)
-            .expect("client event should be serializable");
-
-        trace!("sending event `{}`", any::type_name::<E>());
-        client.send(event_data.channel_id, cursor.into_inner());
-    }
-}
-
-/// Typed version of [`ClientEvent::receive`].
-///
-/// # Safety
-///
-/// The caller must ensure that `events` is [`Events<E>`]
-/// and `event_data` was created for `E`.
-unsafe fn receive<E: Event>(
-    event_data: &ClientEventData,
-    ctx: &mut ServerReceiveCtx,
-    events: PtrMut,
-    server: &mut RepliconServer,
-) {
-    let events: &mut Events<FromClient<E>> = events.deref_mut();
-    for (client_id, message) in server.receive(event_data.channel_id) {
-        let mut cursor = Cursor::new(&*message);
-        match event_data.deserialize::<E>(ctx, &mut cursor) {
-            Ok(event) => {
-                trace!(
-                    "applying event `{}` from `{client_id:?}`",
-                    any::type_name::<E>()
-                );
-                events.send(FromClient { client_id, event });
-            }
-            Err(e) => debug!("unable to deserialize event from {client_id:?}: {e}"),
-        }
-    }
-}
-
-/// Typed version of [`ClientEvent::resend_locally`].
-///
-/// # Safety
-///
-/// The caller must ensure that `events` is [`Events<E>`] and `server_events` is [`Events<ToClients<E>>`].
-unsafe fn resend_locally<E: Event>(client_events: PtrMut, events: PtrMut) {
-    let client_events: &mut Events<FromClient<E>> = client_events.deref_mut();
-    let events: &mut Events<E> = events.deref_mut();
-    client_events.send_batch(events.drain().map(|event| FromClient {
-        client_id: ClientId::SERVER,
-        event,
-    }));
-}
-
-/// Typed version of [`ClientEvent::reset`].
-///
-/// # Safety
-///
-/// The caller must ensure that `events` is [`Events<E>`].
-unsafe fn reset<E: Event>(events: PtrMut) {
-    let events: &mut Events<E> = events.deref_mut();
-    let drained_count = events.drain().count();
-    if drained_count > 0 {
-        warn!("discarded {drained_count} client events due to a disconnect");
-    }
-}
 
 /// Default event serialization function.
 pub fn default_serialize<E: Event + Serialize>(
